Updated readme with a link to the installation
I was trying to run first the build command when I was trying to follow the instructions because that was the first code that I saw and ran into an error. I have to read the whole instruction first to find out that I need to do the installation instruction at the bottom of the document, might help first time users to know the Terminus plugin installation before running any code outlined in this doc.