feat: follow pagination links in collection responses
The calls to the GitHub API may result in a collection of items. When
that happens GitHub may paginate the number of results per request.
If that is the case, there will be pagination links in the response
headers.

This PR will introspect the response to determine if it is for a
paginated resource with more results outside of the current response.
The intent is to accumulate the results for that response so the result
set is complete