allow build:env:create to clone test, live, etc
This commit fixes an error with `build:env:create` which currently depends on the source environment being the `dev` environment (since it is the only one that has a `git_url`).

This is fixed by always using the `dev` environment for `BuildToolsBase ::addPantheonRemote()` since that is the only logical entry point to the codebase of a site. A new branch will still be created for the multidev environment as expected.